Fixed compilation on gcc-3, ia32
git-svn-id: http://yate.null.ro/svn/yate/trunk@872 acf43c95-373e-0410-b603-e72c3f656dc1

+++ contrib/yiax/frame.cpp
@@ -411,7 +411,7 @@ ObjList* IAXFrame::getIEList(const IAXFullFrame* frame, bool& invalid)
 	return 0;
     }
     for (i = 1; i < len;) {
-	if (i + data[i] >= len) {
+	if (i + (unsigned int)data[i] >= len) {
 	    i = 0xFFFF;
 	    break;
 	}
